Created login controller
This file will be used by existing users to access their accounts. The server will likely search for their email address in the relevant table, compare the password with the stored one and, if the credentials match, sign the user into their account, storing their account id as a session variable.