Imported initial database script into version control
This script should not be modified, as that could lead to desyncs if someone has already created their database, and would then need to drop the entire thing. Instead, any changes we make to the DDL will be described in additional .sql files, making it easier for existing databases to be brought up to speed.