Added migration to change primary keys
This will change some of the primary keys to auto-incrementing integers, which will make it a lot easier to add new entries to these tables. Due to foreign key constraints, the customer and seller tables couldn't have their primary keys modified, so I will fix that in a subsequent migration.