Added migration to reverse the relation between sellers and stores
As it turns out, we want the store to belong to a seller, not the other way around. As such, we need to reverse the relation between the seller and the store. I have confirmed that this sequence of migrations will not cause any errors, assuming the database is currently in a valid state as described by the previous migrations.