Fixed database connection issues
As it turns out, the way I had been trying to interpolate the connection parameters was causing issues, since it was trying to use those variables as the names of a variable to use, which it wasn't supposed to do. Additionally, I placed the database creation inside of a try block, in order to avoid sensitive information being printed in the event of a database error.