Replaced logical validation with exception-based validation
Rather than checking an array to see if any errors have occurred, the first missing field will throw an exception, which will cause the server to check each expected field, adding error messages as appropriate. The main difference is that this guarantees that, no matter what, the script will not attempt to continue after it finds a missing field.